add_intercept_column {hardhat}	R Documentation

Add an intercept column to data

Description

This function adds an integer column of 1's to data.

Usage

add_intercept_column(data, name = "(Intercept)", ..., call = current_env())

Arguments

data
A data frame or matrix.

name
The name for the intercept column. Defaults to "(Intercept)", which is the same name that stats::lm() uses.

...
These dots are for future extensions and must be empty.

call
The call used for errors and warnings.

Details

If a column named name already exists in data, then data is returned unchanged and a warning is issued.

Value

data with an intercept column.

Examples

add_intercept_column(mtcars)

add_intercept_column(mtcars, "intercept")

add_intercept_column(as.matrix(mtcars))

[Package hardhat version 1.4.2 Index]