Recover original factor levels

A factor column of new_data might have a subset of the original factor levels for some reason. This function checks for that, and recodes the new factor to have the original levels, with a warning.

enforce_new_data_level_recovery(new_data, original_levels)

Arguments

new_data
A data frame of new predictors and possibly outcomes.

original_levels
A named list of the original levels of either the outcomes or predictors. The names match the factor column names in new_data, and the values are character vectors of the required levels.

Details

If this function is used in your package, it is a good idea to call enforce_new_data_novel_levels() first, as that will take care of any new levels if drop_novel = TRUE. This will ensure that new_data factor columns can only have either exactly the right levels, or a subset of the correct levels.

If enforce_new_data_novel_levels() was not run, or drop_novel = FALSE, then enforce_new_data_level_recovery() follows the following heuristic: if any novel levels are detected, then the novel levels are moved to the end, and the remaining levels are reordered to match the order in original_levels. This only has an affect for ordered factors, but is a warning worth throwing.

Validation

hardhat provides validation functions at three levels.

check_*(): check a condition, and return a list. The list always contains at least one element, ok, a logical that specifies if the check passed. Each check also has check specific elements in the returned list that can be used to construct meaningful error messages.

validate_*(): check a condition, and error if it does not pass. These functions call their corresponding check function, and then provide a default error message. If you, as a developer, want a different error message, then call the check_*() function yourself, and provide your own validation function.

enforce_*(): check a condition, modify the input with a warning if the check fails, and return the modified input. These functions are used when user input is incorrect in some non-critical way, and you can correct it automatically, with a warning.

See also

Examples

# ---------------------------------------------------------------------------
# Setup

iris <- tibble::as_tibble(iris)
train <- iris[1:100,]
test <- iris[101:150,]

# ---------------------------------------------------------------------------
# Use with get_levels()

# If rolling your package, get_levels() can be useful alongside this check
original_levels <- get_levels(train)

# All good!
enforce_new_data_level_recovery(test, original_levels)

#> # A tibble: 50 x 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species  
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>    
#>  1          6.3         3.3          6           2.5 virginica
#>  2          5.8         2.7          5.1         1.9 virginica
#>  3          7.1         3            5.9         2.1 virginica
#>  4          6.3         2.9          5.6         1.8 virginica
#>  5          6.5         3            5.8         2.2 virginica
#>  6          7.6         3            6.6         2.1 virginica
#>  7          4.9         2.5          4.5         1.7 virginica
#>  8          7.3         2.9          6.3         1.8 virginica
#>  9          6.7         2.5          5.8         1.8 virginica
#> 10          7.2         3.6          6.1         2.5 virginica
#> # … with 40 more rows

# ---------------------------------------------------------------------------
# Internally, forge() uses this check like so:

x <- mold(Sepal.Length ~ Species, train)

# No problems here!
enforce_new_data_level_recovery(test, x$engine$info$predictors$levels)

#> # A tibble: 50 x 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species  
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>    
#>  1          6.3         3.3          6           2.5 virginica
#>  2          5.8         2.7          5.1         1.9 virginica
#>  3          7.1         3            5.9         2.1 virginica
#>  4          6.3         2.9          5.6         1.8 virginica
#>  5          6.5         3            5.8         2.2 virginica
#>  6          7.6         3            6.6         2.1 virginica
#>  7          4.9         2.5          4.5         1.7 virginica
#>  8          7.3         2.9          6.3         1.8 virginica
#>  9          6.7         2.5          5.8         1.8 virginica
#> 10          7.2         3.6          6.1         2.5 virginica
#> # … with 40 more rows

# Missing 2 levels
bad_test <- test
bad_test$Species <- droplevels(bad_test$Species)

# Restores the levels with a warning
enforce_new_data_level_recovery(bad_test, x$engine$info$predictors$levels)

#> Warning: The following original factor levels are missing for column, 'Species', and have been restored: 'setosa', 'versicolor'.

#> # A tibble: 50 x 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species  
#>           <dbl>       <dbl>        <dbl>       <dbl> <fct>    
#>  1          6.3         3.3          6           2.5 virginica
#>  2          5.8         2.7          5.1         1.9 virginica
#>  3          7.1         3            5.9         2.1 virginica
#>  4          6.3         2.9          5.6         1.8 virginica
#>  5          6.5         3            5.8         2.2 virginica
#>  6          7.6         3            6.6         2.1 virginica
#>  7          4.9         2.5          4.5         1.7 virginica
#>  8          7.3         2.9          6.3         1.8 virginica
#>  9          6.7         2.5          5.8         1.8 virginica
#> 10          7.2         3.6          6.1         2.5 virginica
#> # … with 40 more rows

# ---------------------------------------------------------------------------
# Novel levels and ordered factors

# Ideally, one would run enforce_new_data_novel_levels() before
# running the level recovery function to catch novel levels. If
# this is not the case, then the original levels take priority
# and then any "novel" levels are appended to the end in the order
# they appear

# df2 is:
# - missing level 'b'
# - has an additional level, 'd'
df1 <- data.frame(x = ordered(c("a", "b", "c")))
df2 <- data.frame(x = ordered(c("d", "c", "a"), levels = c("d", "c", "a")))

# 'b' is recovered with a warning
df2_recovered <- enforce_new_data_level_recovery(df2, get_levels(df1))

#> Warning: The following original factor levels are missing for column, 'x', and have been restored: 'b'.

#> Warning: The following novel levels were detected in 'x': 'd'. The order of the `original_levels` has been restored, and the novel levels have been moved to the end.

# 'd' is kept around, but is shifted to the end as the
# order of the original levels takes priority
df2_recovered$x

#> [1] d c a
#> Levels: a < b < c < d