Forge - Formula Method

For the terms preprocessor, forge() does the following:

Calls shrink() to trim new_data to only the required columns and coerce new_data to a tibble.

Calls scream() to perform validation on the structure of the columns of new_data.

Predictors

Runs stats::model.frame() on new_data using the stored terms object corresponding to the predictors.

If, in the original mold() call, indicators = TRUE was set, it then runs stats::model.matrix() on the result.

If any offsets are present from using offset() in the original call to mold(), then they are extracted with model_offset().

If intercept = TRUE in the original call to mold(), then an intecept column is added.

Coerces the result of the above steps to a tibble.

Outcomes

Runs stats::model.frame() on new_data using the stored terms object corresponding to the outcomes.

Coerces the result to a tibble.

# S3 method for terms_preprocessor
forge(preprocessor, new_data,
  outcomes = FALSE, ...)

Arguments

preprocessor
A "terms_preprocessor".

new_data
A data frame or matrix to preprocess.

outcomes
A logical. Should the outcomes be processed and returned as well?

...
Not currently used.

Value

A named list with 3 elements:

predictors: A tibble containing the preprocessed new_data predictors.

outcomes: If outcomes = TRUE, a tibble containing the preprocessed new_data outcomes. Otherwise, NULL.

offset: If the preprocessor was a "terms_preprocessor", and offsets were specified in the formula, this is a tibble containing the preprocessed offsets. Otherwise, NULL.

Examples

# ---------------------------------------------------------------------------
# Setup

train <- iris[1:100,]
test <- iris[101:150,]

# ---------------------------------------------------------------------------
# Formula Example

# Call mold() with the training data
processed <- mold(
  log(Sepal.Length) ~ Sepal.Length + Species,
  train,
  intercept = TRUE
)

# Then, call forge() with the preprocessor and the test data
# to have it preprocess the test data in the same way
forge(processed$preprocessor, test)

#> $predictors
#> # A tibble: 50 x 4
#>    `(Intercept)` Sepal.Length Speciesversicolor Speciesvirginica
#>            <dbl>        <dbl>             <dbl>            <dbl>
#>  1             1          6.3                 0                1
#>  2             1          5.8                 0                1
#>  3             1          7.1                 0                1
#>  4             1          6.3                 0                1
#>  5             1          6.5                 0                1
#>  6             1          7.6                 0                1
#>  7             1          4.9                 0                1
#>  8             1          7.3                 0                1
#>  9             1          6.7                 0                1
#> 10             1          7.2                 0                1
#> # … with 40 more rows
#> 
#> $outcomes
#> NULL
#> 
#> $offset
#> NULL
#> 

# Use `outcomes = TRUE` to also extract the preprocessed outcome
forge(processed$preprocessor, test, outcomes = TRUE)

#> $predictors
#> # A tibble: 50 x 4
#>    `(Intercept)` Sepal.Length Speciesversicolor Speciesvirginica
#>            <dbl>        <dbl>             <dbl>            <dbl>
#>  1             1          6.3                 0                1
#>  2             1          5.8                 0                1
#>  3             1          7.1                 0                1
#>  4             1          6.3                 0                1
#>  5             1          6.5                 0                1
#>  6             1          7.6                 0                1
#>  7             1          4.9                 0                1
#>  8             1          7.3                 0                1
#>  9             1          6.7                 0                1
#> 10             1          7.2                 0                1
#> # … with 40 more rows
#> 
#> $outcomes
#> # A tibble: 50 x 1
#>    `log(Sepal.Length)`
#>                  <dbl>
#>  1                1.84
#>  2                1.76
#>  3                1.96
#>  4                1.84
#>  5                1.87
#>  6                2.03
#>  7                1.59
#>  8                1.99
#>  9                1.90
#> 10                1.97
#> # … with 40 more rows
#> 
#> $offset
#> NULL
#> 

# ---------------------------------------------------------------------------
# Dummy variables

# If factors are not expanded in mold()...
processed <- mold(Sepal.Width ~ Species, train, indicators = FALSE)

# ...then they aren't expanded in forge() either
forge(processed$preprocessor, test)

#> $predictors
#> # A tibble: 50 x 1
#>    Species  
#>    <fct>    
#>  1 virginica
#>  2 virginica
#>  3 virginica
#>  4 virginica
#>  5 virginica
#>  6 virginica
#>  7 virginica
#>  8 virginica
#>  9 virginica
#> 10 virginica
#> # … with 40 more rows
#> 
#> $outcomes
#> NULL
#> 
#> $offset
#> NULL
#> 

# ---------------------------------------------------------------------------
# Multivariate outcomes

# Multivariate formulas specified in mold()
# carry over into forge()
processed <- mold(Sepal.Width + log(Sepal.Length) ~ Species, train)

forge(processed$preprocessor, test, outcomes = TRUE)

#> $predictors
#> # A tibble: 50 x 3
#>    Speciessetosa Speciesversicolor Speciesvirginica
#>            <dbl>             <dbl>            <dbl>
#>  1             0                 0                1
#>  2             0                 0                1
#>  3             0                 0                1
#>  4             0                 0                1
#>  5             0                 0                1
#>  6             0                 0                1
#>  7             0                 0                1
#>  8             0                 0                1
#>  9             0                 0                1
#> 10             0                 0                1
#> # … with 40 more rows
#> 
#> $outcomes
#> # A tibble: 50 x 2
#>    Sepal.Width `log(Sepal.Length)`
#>          <dbl>               <dbl>
#>  1         3.3                1.84
#>  2         2.7                1.76
#>  3         3                  1.96
#>  4         2.9                1.84
#>  5         3                  1.87
#>  6         3                  2.03
#>  7         2.5                1.59
#>  8         2.9                1.99
#>  9         2.5                1.90
#> 10         3.6                1.97
#> # … with 40 more rows
#> 
#> $offset
#> NULL
#> 

# ---------------------------------------------------------------------------
# Offsets

# Offsets specified in mold() are computed in forge() as well,
# and are placed in the `$offset` slot of the result
processed <- mold(
  Sepal.Width ~ Species + offset(Sepal.Length) + offset(Petal.Width),
  train
)

forge(processed$preprocessor, test)

#> $predictors
#> # A tibble: 50 x 3
#>    Speciessetosa Speciesversicolor Speciesvirginica
#>            <dbl>             <dbl>            <dbl>
#>  1             0                 0                1
#>  2             0                 0                1
#>  3             0                 0                1
#>  4             0                 0                1
#>  5             0                 0                1
#>  6             0                 0                1
#>  7             0                 0                1
#>  8             0                 0                1
#>  9             0                 0                1
#> 10             0                 0                1
#> # … with 40 more rows
#> 
#> $outcomes
#> NULL
#> 
#> $offset
#> # A tibble: 50 x 1
#>    .offset
#>      <dbl>
#>  1     8.8
#>  2     7.7
#>  3     9.2
#>  4     8.1
#>  5     8.7
#>  6     9.7
#>  7     6.6
#>  8     9.1
#>  9     8.5
#> 10     9.7
#> # … with 40 more rows
#>