Forge - XY Method

For the default preprocessor, forge() does the following:

Calls shrink() to trim new_data to only the required columns and coerce new_data to a tibble.

Calls scream() to perform validation on the structure of the columns of new_data.

Calls on the default preprocessor engine to potentially add an intercept column onto new_data, if the corresponding call to mold() used one.

# S3 method for default_preprocessor
forge(preprocessor, new_data,
  outcomes = FALSE, ...)

Arguments

preprocessor
A "default_preprocessor".

new_data
A data frame or matrix to preprocess.

outcomes
A logical. Should the outcomes be processed and returned as well?

...
Not currently used.

Value

A named list with 3 elements:

predictors: A tibble containing the preprocessed new_data predictors.

outcomes: If outcomes = TRUE, a tibble containing the preprocessed new_data outcomes. Otherwise, NULL.

offset: If the preprocessor was a "terms_preprocessor", and offsets were specified in the formula, this is a tibble containing the preprocessed offsets. Otherwise, NULL.

Details

The one special thing about the XY method of forge() is the behavior of outcomes = TRUE when a vector y value was provided to the original call to mold() (which generated the preprocessor). In that case, mold() converted y into a tibble, with a default name of .outcome. This is the column that forge() will look for in new_data to preprocess. See the examples section for a demonstration of this.

Examples

# ---------------------------------------------------------------------------
# Setup

train <- iris[1:100,]
test <- iris[101:150,]

train_x <- train[, "Sepal.Length", drop = FALSE]
train_y <- train[, "Species", drop = FALSE]

test_x <- test[, "Sepal.Length", drop = FALSE]
test_y <- test[, "Species", drop = FALSE]

# ---------------------------------------------------------------------------
# XY Example

# First, call mold() with the training data
processed <- mold(train_x, train_y)

# Then, call forge() with the preprocessor and the test data
# to have it preprocess the test data in the same way
forge(processed$preprocessor, test_x)

#> $predictors
#> # A tibble: 50 x 1
#>    Sepal.Length
#>           <dbl>
#>  1          6.3
#>  2          5.8
#>  3          7.1
#>  4          6.3
#>  5          6.5
#>  6          7.6
#>  7          4.9
#>  8          7.3
#>  9          6.7
#> 10          7.2
#> # … with 40 more rows
#> 
#> $outcomes
#> NULL
#> 
#> $offset
#> NULL
#> 

# ---------------------------------------------------------------------------
# Intercept

processed <- mold(train_x, train_y, intercept = TRUE)

forge(processed$preprocessor, test_x)

#> $predictors
#> # A tibble: 50 x 2
#>    `(Intercept)` Sepal.Length
#>            <int>        <dbl>
#>  1             1          6.3
#>  2             1          5.8
#>  3             1          7.1
#>  4             1          6.3
#>  5             1          6.5
#>  6             1          7.6
#>  7             1          4.9
#>  8             1          7.3
#>  9             1          6.7
#> 10             1          7.2
#> # … with 40 more rows
#> 
#> $outcomes
#> NULL
#> 
#> $offset
#> NULL
#> 

# ---------------------------------------------------------------------------
# XY Method and forge(outcomes = TRUE)

# You can request that the new outcome columns are preprocessed as well, but
# they have to be present in `new_data`!

processed <- mold(train_x, train_y)

# Can't do this!
# forge(processed$preprocessor, test_x, outcomes = TRUE)

# Need to use the full test set, including `y`
forge(processed$preprocessor, test, outcomes = TRUE)

#> $predictors
#> # A tibble: 50 x 1
#>    Sepal.Length
#>           <dbl>
#>  1          6.3
#>  2          5.8
#>  3          7.1
#>  4          6.3
#>  5          6.5
#>  6          7.6
#>  7          4.9
#>  8          7.3
#>  9          6.7
#> 10          7.2
#> # … with 40 more rows
#> 
#> $outcomes
#> # A tibble: 50 x 1
#>    Species  
#>    <fct>    
#>  1 virginica
#>  2 virginica
#>  3 virginica
#>  4 virginica
#>  5 virginica
#>  6 virginica
#>  7 virginica
#>  8 virginica
#>  9 virginica
#> 10 virginica
#> # … with 40 more rows
#> 
#> $offset
#> NULL
#> 

# With the XY method, if the Y value used in `mold()` is a vector,
# then a column name of `.outcome` is automatically generated.
# This name is what forge() looks for in `new_data`.

# Y is a vector!
y_vec <- train_y$Species

processed_vec <- mold(train_x, y_vec)

# This throws an informative error that tell you
# to include an `".outcome"` column in `new_data`.
# NOT RUN {
forge(processed_vec$preprocessor, iris, outcomes = TRUE)
# }

test2 <- test
test2$.outcome <- test2$Species
test2$Species <- NULL

# This works, and returns a tibble in the $outcomes slot
forge(processed_vec$preprocessor, test2, outcomes = TRUE)

#> $predictors
#> # A tibble: 50 x 1
#>    Sepal.Length
#>           <dbl>
#>  1          6.3
#>  2          5.8
#>  3          7.1
#>  4          6.3
#>  5          6.5
#>  6          7.6
#>  7          4.9
#>  8          7.3
#>  9          6.7
#> 10          7.2
#> # … with 40 more rows
#> 
#> $outcomes
#> # A tibble: 50 x 1
#>    .outcome 
#>    <fct>    
#>  1 virginica
#>  2 virginica
#>  3 virginica
#>  4 virginica
#>  5 virginica
#>  6 virginica
#>  7 virginica
#>  8 virginica
#>  9 virginica
#> 10 virginica
#> # … with 40 more rows
#> 
#> $offset
#> NULL
#>