get_data_classes {hardhat}	R Documentation

Extract data classes from a data frame or matrix

Description

When predicting from a model, it is often important for the new_data to have the same classes as the original data used to fit the model. get_data_classes() extracts the classes from the original training data.

Usage

get_data_classes(data, ..., call = current_env())

Arguments

data
A data frame or matrix.

...
These dots are for future extensions and must be empty.

call
The call used for errors and warnings.

Value

A named list. The names are the column names of data and the values are character vectors containing the class of that column.

Examples

get_data_classes(iris)

get_data_classes(as.matrix(mtcars))

# Unlike .MFclass(), the full class
# vector is returned
data <- data.frame(col = ordered(c("a", "b")))

.MFclass(data$col)

get_data_classes(data)

[Package hardhat version 1.4.2 Index]