hardhat-package {hardhat}	R Documentation

hardhat: Construct Modeling Packages

Description

Building modeling packages is hard. A large amount of effort generally goes into providing an implementation for a new method that is efficient, fast, and correct, but often less emphasis is put on the user interface. A good interface requires specialized knowledge about S3 methods and formulas, which the average package developer might not have. The goal of 'hardhat' is to reduce the burden around building new modeling packages by providing functionality for preprocessing, predicting, and validating input.

Author(s)

Maintainer: Hannah Frick hannah@posit.co (ORCID)

Authors:

Davis Vaughan davis@posit.co

Max Kuhn max@posit.co

Other contributors:

Posit Software, PBC (03wc8by49) [copyright holder, funder]

See Also

Useful links:

https://github.com/tidymodels/hardhat

https://hardhat.tidymodels.org

Report bugs at https://github.com/tidymodels/hardhat/issues

[Package hardhat version 1.4.2 Index]