mold {hardhat}	R Documentation

Mold data for modeling

Description

mold() applies the appropriate processing steps required to get training data ready to be fed into a model. It does this through the use of various blueprints that understand how to preprocess data that come in various forms, such as a formula or a recipe.

All blueprints have consistent return values with the others, but each is unique enough to have its own help page. Click through below to learn how to use each one in conjunction with mold().

XY Method - default_xy_blueprint()

Formula Method - default_formula_blueprint()

Recipes Method - default_recipe_blueprint()

Usage

mold(x, ...)

Arguments

x
An object. See the method specific implementations linked in the Description for more information.

...
Not used.

Value

A named list containing 4 elements:

predictors: A tibble containing the molded predictors to be used in the model.

outcomes: A tibble containing the molded outcomes to be used in the model.

blueprint: A method specific "hardhat_blueprint" object for use when making predictions.

extras: Either NULL if the blueprint returns no extra information, or a named list containing the extra information.

Examples

# See the method specific documentation linked in Description
# for the details of each blueprint, and more examples.

# XY
mold(iris["Sepal.Width"], iris$Species)

# Formula
mold(Species ~ Sepal.Width, iris)

# Recipe
library(recipes)
mold(recipe(Species ~ Sepal.Width, iris), iris)

[Package hardhat version 1.4.2 Index]