Mold - Formula Method

For a formula, mold() does the following:

Predictors

The RHS of the formula is isolated, and converted to its own 1 sided formula: ~ RHS.

Runs stats::model.frame() on the RHS formula and uses data.

If indicators = TRUE, it then runs stats::model.matrix() on the result.

If indicators = FALSE, it does not expand factors with model.matrix(), which also means that interactions are not expanded.

If any offsets are present from using offset(), then they are extracted with model_offset().

If intercept = TRUE, adds an intercept column.

Coerces the result of the above steps to a tibble.

Outcomes

The LHS of the formula is isolated, and converted to its own 1 sided formula: ~ LHS.

Runs stats::model.frame() on the LHS formula and uses data.

Coerces the result of the above steps to a tibble.

# S3 method for formula
mold(formula, data, intercept = FALSE,
  indicators = TRUE, ...)

Arguments

formula
A formula specifying the terms of the model, with the outcomes on the left-hand side of the formula, and the predictors on the right-hand side.

data
A data frame containing the predictors and the outcomes.

intercept
A single logical specifying whether or not to include an intercept in the molded predictors.

indicators
Should factors and interactions be expanded (In other words, should stats::model.matrix() be run)? If FALSE, factor columns are returned without being expanded into dummy variables and a warning is thrown if any interactions are detected.

...
Currently unused.

Value

A named list containing 4 elements:

predictors: A tibble containing the molded predictors to be used in the model.

outcome: A tibble containing the molded outcomes to be used in the model.

preprocessor: A "preprocessor" object for use when making predictions.

offset: A tibble with a single column named ".offset" if an offset was specified in the formula method. Otherwise, NULL.

Details

While not different from base R, the behavior of expanding factors into dummy variables when an intercept is not present should be documented.

When an intercept is present, factors are expanded into K-1 new columns, where K is the number of levels in the factor.

When an intercept is not present, factors are expanded into all K columns (one-hot encoding).

Offsets can be included in the formula method through the use of the inline function stats::offset(). These are returned as a tibble with 1 column named ".offset" in the $outcome slot of the return value.

Differences From Base R

There are a number of differences from base R regarding how formulas are processed by mold() that require some explanation.

Multivariate outcomes can be specified on the LHS using syntax that is similar to the RHS (i.e. outcome_1 + outcome_2 ~ predictors). If any complex calculations are done on the LHS and they return matrices (like stats::poly()), then those matrices are flattened into multiple columns of the tibble after the call to model.frame(). While this is possible, it is not recommended, and if a large amount of preprocessing is required on the outcomes you are better off using a recipes::recipe().

Global variables are not allowed in the formula. An error will be thrown if they are included. All terms in the formula should come from data.

By default, intercepts are not included in the predictor output from the formula. To include an intercept, set intercept = TRUE. Having an intercept argument is consistent with the other mold() methods. More importantly, there are often modeling packages where an intercept is either always or never allowed (for example, the earth package), and they do some fancy footwork to keep the user from providing or removing an intercept. This interface standardizes all of that flexibility in one place.

Examples

# ---------------------------------------------------------------------------
# Formula example

processed <- mold(Sepal.Width ~ Species, iris, intercept = TRUE)

processed$predictors

#> # A tibble: 150 x 3
#>    `(Intercept)` Speciesversicolor Speciesvirginica
#>            <dbl>             <dbl>            <dbl>
#>  1             1                 0                0
#>  2             1                 0                0
#>  3             1                 0                0
#>  4             1                 0                0
#>  5             1                 0                0
#>  6             1                 0                0
#>  7             1                 0                0
#>  8             1                 0                0
#>  9             1                 0                0
#> 10             1                 0                0
#> # … with 140 more rows

processed$outcomes

#> # A tibble: 150 x 1
#>    Sepal.Width
#>          <dbl>
#>  1         3.5
#>  2         3  
#>  3         3.2
#>  4         3.1
#>  5         3.6
#>  6         3.9
#>  7         3.4
#>  8         3.4
#>  9         2.9
#> 10         3.1
#> # … with 140 more rows

# ---------------------------------------------------------------------------
# Factors without an intercept

# No intercept is added
processed <- mold(Sepal.Width ~ Species, iris)

# So factor columns are completely expanded
# into all `K` columns (the number of levels)
processed$predictors

#> # A tibble: 150 x 3
#>    Speciessetosa Speciesversicolor Speciesvirginica
#>            <dbl>             <dbl>            <dbl>
#>  1             1                 0                0
#>  2             1                 0                0
#>  3             1                 0                0
#>  4             1                 0                0
#>  5             1                 0                0
#>  6             1                 0                0
#>  7             1                 0                0
#>  8             1                 0                0
#>  9             1                 0                0
#> 10             1                 0                0
#> # … with 140 more rows

# ---------------------------------------------------------------------------
# Global variables

y <- rep(1, times = nrow(iris))

# In base R, global variables are allowed in a model formula
frame <- model.frame(Species ~ y + Sepal.Length, iris)
head(frame)

#>   Species y Sepal.Length
#> 1  setosa 1          5.1
#> 2  setosa 1          4.9
#> 3  setosa 1          4.7
#> 4  setosa 1          4.6
#> 5  setosa 1          5.0
#> 6  setosa 1          5.4

# mold() does not allow them, and throws an error
tryCatch(
  expr = mold(Species ~ y + Sepal.Length, iris),
  error = function(e) print(e$message)
)

#> [1] "object 'y' not found"

# ---------------------------------------------------------------------------
# Dummy variables and interactions

# By default, factor columns are expanded
# And interactions are created, both by
# calling model.matrix(). Some models (like
# tree based models) can take factors directly
# but still might want to use the formula method.
# In those cases, set `indicators = FALSE` to not
# run model.matrix()

processed <- mold(Sepal.Width ~ Species, iris, indicators = FALSE)

processed$predictors

#> # A tibble: 150 x 1
#>    Species
#>    <fct>  
#>  1 setosa 
#>  2 setosa 
#>  3 setosa 
#>  4 setosa 
#>  5 setosa 
#>  6 setosa 
#>  7 setosa 
#>  8 setosa 
#>  9 setosa 
#> 10 setosa 
#> # … with 140 more rows

# No interactions are expanded when `indicators = FALSE`, but
# the terms specified in the interaction show up in the predictors
# tibble, with a warning

# NOT RUN {
mold(Sepal.Width ~ Sepal.Length:Species, iris, indicators = FALSE)
# }

# ---------------------------------------------------------------------------
# Multivariate outcomes

# Multivariate formulas can be specified easily
processed <- mold(Sepal.Width + log(Sepal.Length) ~ Species, iris)
processed$outcomes

#> # A tibble: 150 x 2
#>    Sepal.Width `log(Sepal.Length)`
#>          <dbl>               <dbl>
#>  1         3.5                1.63
#>  2         3                  1.59
#>  3         3.2                1.55
#>  4         3.1                1.53
#>  5         3.6                1.61
#>  6         3.9                1.69
#>  7         3.4                1.53
#>  8         3.4                1.61
#>  9         2.9                1.48
#> 10         3.1                1.59
#> # … with 140 more rows

# Inline functions on the LHS are run, but any matrix
# output is flattened (like what happens in `model.matrix()`)
# (essentially this means you don't wind up with columns
# in the tibble that are matrices)
processed <- mold(poly(Sepal.Length, degree = 2) ~ Species, iris)
processed$outcomes

#> # A tibble: 150 x 2
#>    `poly(Sepal.Length, degree = 2).1` `poly(Sepal.Length, degree = 2).2`
#>                                 <dbl>                              <dbl>
#>  1                            -0.0735                            0.00652
#>  2                            -0.0933                            0.0469 
#>  3                            -0.113                             0.0957 
#>  4                            -0.123                             0.123  
#>  5                            -0.0834                            0.0257 
#>  6                            -0.0439                           -0.0385 
#>  7                            -0.123                             0.123  
#>  8                            -0.0834                            0.0257 
#>  9                            -0.143                             0.184  
#> 10                            -0.0933                            0.0469 
#> # … with 140 more rows

# TRUE
ncol(processed$outcomes) == 2

#> [1] TRUE

# ---------------------------------------------------------------------------
# Offsets

# Offsets are handled specially in base R, so they deserve special
# treatment here as well. You can add offsets using the inline function
# offset()
processed <- mold(Sepal.Width ~ offset(Sepal.Length) + Species, iris)

processed$offset

#> # A tibble: 150 x 1
#>    .offset
#>      <dbl>
#>  1     5.1
#>  2     4.9
#>  3     4.7
#>  4     4.6
#>  5     5  
#>  6     5.4
#>  7     4.6
#>  8     5  
#>  9     4.4
#> 10     4.9
#> # … with 140 more rows

# Multiple offsets can be included, and they get added together
processed <- mold(
  Sepal.Width ~ offset(Sepal.Length) + offset(Petal.Width),
  iris
)

identical(
  processed$offset$.offset,
  iris$Sepal.Length + iris$Petal.Width
)

#> [1] TRUE