Mold - Recipes Method

For a recipe, mold() does the following:

Calls recipes::prep() to prep the recipe.

Calls recipes::juice() to extract the outcomes and predictors. These are returned as tibbles.

If intercept = TRUE, adds an intercept column to the predictors.

# S3 method for recipe
mold(x, data, intercept = FALSE, ...)

Arguments

x
An unprepped recipe created from recipes::recipe().

data
A data frame containing the predictors and the outcomes.

intercept
A single logical specifying whether or not to include an intercept in the molded predictors.

...
Currently unused.

Value

A named list containing 4 elements:

predictors: A tibble containing the molded predictors to be used in the model.

outcome: A tibble containing the molded outcomes to be used in the model.

preprocessor: A "preprocessor" object for use when making predictions.

offset: A tibble with a single column named ".offset" if an offset was specified in the formula method. Otherwise, NULL.

Examples

library(recipes)

# ---------------------------------------------------------------------------
# Recipes example

# Create a recipe that logs a predictor
rec <- recipe(Species ~ Sepal.Length + Sepal.Width, iris) %>%
   step_log(Sepal.Length)

processed <- mold(rec, iris)

# Sepal.Length has been logged
processed$predictors

#> # A tibble: 150 x 2
#>    Sepal.Length Sepal.Width
#>           <dbl>       <dbl>
#>  1         1.63         3.5
#>  2         1.59         3  
#>  3         1.55         3.2
#>  4         1.53         3.1
#>  5         1.61         3.6
#>  6         1.69         3.9
#>  7         1.53         3.4
#>  8         1.61         3.4
#>  9         1.48         2.9
#> 10         1.59         3.1
#> # … with 140 more rows

processed$outcomes

#> # A tibble: 150 x 1
#>    Species
#>    <fct>  
#>  1 setosa 
#>  2 setosa 
#>  3 setosa 
#>  4 setosa 
#>  5 setosa 
#>  6 setosa 
#>  7 setosa 
#>  8 setosa 
#>  9 setosa 
#> 10 setosa 
#> # … with 140 more rows

# The underlying engine is a prepped recipe
processed$preprocessor$engine

#> Data Recipe
#> 
#> Inputs:
#> 
#>       role #variables
#>    outcome          1
#>  predictor          2
#> 
#> Training data contained 150 data points and no missing data.
#> 
#> Operations:
#> 
#> Log transformation on Sepal.Length [trained]

# ---------------------------------------------------------------------------
# With an intercept

# You can add a predictor with `intercept = TRUE`
processed <- mold(rec, iris, intercept = TRUE)

processed$predictors

#> # A tibble: 150 x 3
#>    `(Intercept)` Sepal.Length Sepal.Width
#>            <int>        <dbl>       <dbl>
#>  1             1         1.63         3.5
#>  2             1         1.59         3  
#>  3             1         1.55         3.2
#>  4             1         1.53         3.1
#>  5             1         1.61         3.6
#>  6             1         1.69         3.9
#>  7             1         1.53         3.4
#>  8             1         1.61         3.4
#>  9             1         1.48         2.9
#> 10             1         1.59         3.1
#> # … with 140 more rows

# But you also could have used a recipe step
rec2 <- step_intercept(rec)

mold(rec2, iris)$predictors

#> # A tibble: 150 x 3
#>    intercept Sepal.Length Sepal.Width
#>        <dbl>        <dbl>       <dbl>
#>  1         1         1.63         3.5
#>  2         1         1.59         3  
#>  3         1         1.55         3.2
#>  4         1         1.53         3.1
#>  5         1         1.61         3.6
#>  6         1         1.69         3.9
#>  7         1         1.53         3.4
#>  8         1         1.61         3.4
#>  9         1         1.48         2.9
#> 10         1         1.59         3.1
#> # … with 140 more rows