Mold - XY Method

For a data frame / matrix, mold() does the following:

Converts x to a tibble.

Adds an intercept column to x if intercept = TRUE.

Runs standardize() on y.

# S3 method for data.frame
mold(x, y, intercept = FALSE, ...)

# S3 method for matrix
mold(x, y, intercept = FALSE, ...)

Arguments

x
A data frame or matrix containing the predictors.

y
A data frame, matrix, or vector containing the outcome(s).

intercept
A single logical specifying whether or not to include an intercept in the molded predictors.

...
Currently unused.

Value

A named list containing 4 elements:

predictors: A tibble containing the molded predictors to be used in the model.

outcome: A tibble containing the molded outcomes to be used in the model.

preprocessor: A "preprocessor" object for use when making predictions.

offset: A tibble with a single column named ".offset" if an offset was specified in the formula method. Otherwise, NULL.

Details

As documented in standardize(), if y is a vector, then the returned outcomes tibble has 1 column with a standardized name of ".outcome".

Examples

# ---------------------------------------------------------------------------
# XY Example

x <- iris[, c("Sepal.Width", "Species"), drop = FALSE]
y <- iris[, "Sepal.Length", drop = FALSE]

processed <- mold(x, y)

# The predictors are returned as a tibble
processed$predictors

#> # A tibble: 150 x 2
#>    Sepal.Width Species
#>          <dbl> <fct>  
#>  1         3.5 setosa 
#>  2         3   setosa 
#>  3         3.2 setosa 
#>  4         3.1 setosa 
#>  5         3.6 setosa 
#>  6         3.9 setosa 
#>  7         3.4 setosa 
#>  8         3.4 setosa 
#>  9         2.9 setosa 
#> 10         3.1 setosa 
#> # … with 140 more rows

# So are the outcomes
processed$outcomes

#> # A tibble: 150 x 1
#>    Sepal.Length
#>           <dbl>
#>  1          5.1
#>  2          4.9
#>  3          4.7
#>  4          4.6
#>  5          5  
#>  6          5.4
#>  7          4.6
#>  8          5  
#>  9          4.4
#> 10          4.9
#> # … with 140 more rows

# A default preprocessor is also returned
# This contains all of the information required
# to preprocess new data at prediction time.
processed$preprocessor

#> Default Preprocessor: 
#>  
#> # Predictors: 2 
#>   # Outcomes: 1 
#>    Intercept: FALSE 

# The preprocessor should be stored in the model object
# and is later used by forge()
forge(processed$preprocessor, iris)

#> $predictors
#> # A tibble: 150 x 2
#>    Sepal.Width Species
#>          <dbl> <fct>  
#>  1         3.5 setosa 
#>  2         3   setosa 
#>  3         3.2 setosa 
#>  4         3.1 setosa 
#>  5         3.6 setosa 
#>  6         3.9 setosa 
#>  7         3.4 setosa 
#>  8         3.4 setosa 
#>  9         2.9 setosa 
#> 10         3.1 setosa 
#> # … with 140 more rows
#> 
#> $outcomes
#> NULL
#> 
#> $offset
#> NULL
#> 

# ---------------------------------------------------------------------------
# Y as a vector

# Often, `y` will be supplied as a vector to a model. `mold()` handles
# this by letting `standardize()` convert `y` to a tibble, adding the
# default column name, `".outcome"`.
y_vec <- y$Sepal.Length

mold(x, y_vec)$outcomes

#> # A tibble: 150 x 1
#>    .outcome
#>       <dbl>
#>  1      5.1
#>  2      4.9
#>  3      4.7
#>  4      4.6
#>  5      5  
#>  6      5.4
#>  7      4.6
#>  8      5  
#>  9      4.4
#> 10      4.9
#> # … with 140 more rows