Create a new preprocessing engine

These are the base classes for creating new preprocessing engines. All engines inherit from the one created by new_engine(), and the default method specific engines inherit from the other three here.

If you want to create your own processing engine for a specific method, generally you will subclass one of the method specific engines here. If you want to create a completely new preprocessing engine for a totally new preprocessing method (i.e. not the formula, xy, or recipe method) then you should subclass new_engine().

new_formula_engine(mold, forge, intercept = FALSE, ptypes = NULL,
  formula = NULL, indicators = TRUE, ..., subclass = character())

new_recipe_engine(mold, forge, intercept = FALSE, ptypes = NULL,
  recipe = NULL, ..., subclass = character())

new_xy_engine(mold, forge, intercept = FALSE, ptypes = NULL, ...,
  subclass = character())

new_engine(mold, forge, intercept = FALSE, ptypes = NULL, ...,
  subclass = character())

Arguments

mold
A named list with two elements, clean and process, see the new_engine() section, Mold Functions, for details.

forge
A named list with two elements, clean and process, see the new_engine() section, Forge Functions, for details.

intercept
A logical. Should an intercept be included in the processed data? This information is used by the process function in the mold and forge function list.

ptypes
Either NULL, or a named list with 2 elements, predictors and outcomes, both of which are 0-row tibbles. ptypes is generated automatically at mold() time and is used to validate new_data at prediction time. At mold() time, the information found in engine$mold$process()$ptype is used to set ptypes for the engine.

formula
Either NULL, or a formula that specifies how the predictors and outcomes should be preprocessed. This argument is set automatically at mold() time.

indicators
A logical. Should factors be expanded into dummy variables?

...
Name-value pairs for additional elements of engines that subclass this engine.

subclass
A character vector. The subclasses of this engine.

recipe
Either NULL, or an unprepped recipe. This argument is set automatically at mold() time.

Value

A preprocessing engine, which is a list containing the inputs used as arguments to the function, along with a class specific to the type of engine being created.

Mold Functions

engine$mold should be a named list with two elements, both of which are functions:

clean: A function that performs initial cleaning of the user's input data to be used in the model.

Arguments:

If this is an xy engine, engine, x and y.

Otherwise, engine and data.

Output: A named list of three elements:

engine: The engine, returned and potentially updated.

If using an xy engine:

x: The cleaned predictor data.

y: The cleaned outcome data.

If not using an xy engine:

data: The cleaned data.

process: A function that performs the actual preprocessing of the data.

Arguments:

If this is an xy engine, engine, x and y.

Otherwise, engine and data.

Output: A named list of 5 elements:

engine: The engine, returned and potentially updated.

predictors: A tibble of predictors.

outcomes: A tibble of outcomes.

ptypes: A named list with 2 elements, predictors and outcomes, where both elements are 0-row tibbles.

extras: Varies based on the engine. If the engine has no extra information, NULL. Otherwise a named list of the extra elements returned by the engine.

Both engine$mold$clean() and engine$mold$process() will be called, in order, from mold().

Forge Functions

engine$forge should be a named list with two elements, both of which are functions:

clean: A function that performs initial cleaning of new_data:

Arguments:

engine, new_data, and outcomes.

Output: A named list of the following elements:

engine: The engine, returned and potentially updated.

predictors: A tibble containing the cleaned predictors.

outcomes: A tibble containing the cleaned outcomes.

extras: A named list of any extras obtained while cleaning. These are passed on to the process() function for further use.

process: A function that performs the actual preprocessing of the data using the known information in the engine.

Arguments:

engine, new_data, outcomes, extras.

Output: A named list of the following elements:

engine: The engine, returned and potentially updated.

predictors: A tibble of the predictors.

outcomes: A tibble of the outcomes, or NULL.

extras: Varies based on the engine. If the engine has no extra information, NULL. Otherwise a named list of the extra elements returned by the engine.

Both engine$forge$clean() and engine$forge$process() will be called, in order, from forge().