new_case_weights {hardhat}	R Documentation

Extend case weights

Description

new_case_weights() is a developer oriented function for constructing a new case weights type. The <case_weights> type itself is an abstract type with very little functionality. Because of this, class is a required argument.

Usage

new_case_weights(x, ..., class)

Arguments

x
An integer or double vector.

...
Name-value pairs defining attributes

class
Name of subclass.

Value

A new subclassed case weights vector.

Examples

new_case_weights(1:5, class = "my_weights")

[Package hardhat version 1.4.2 Index]