😱 Scream.

scream() performs a number of validation checks on new_data, and yells loudly if anything is wrong. scream() performs the following validation:

Checks that the class of each required predictor in new_data is the same as the class used during training.

Checks that all new_data factor columns don't have any new levels when compared with the original data used in training. If there are new levels, they are replaced with NA values and a warning is issued.

Checks that all new_data factor columns aren't missing any factor levels when compared with the original data used in training. If there are missing levels (or misordered levels for ordered factors), then they are restored and a warning is issued.

scream(preprocessor, new_data, outcomes = FALSE)

Arguments

preprocessor
A "preprocessor".

new_data
A data frame containing the new data to check the structure of. This should contain the predictors, and potentially the outcomes if outcomes = TRUE.

outcomes
A logical. Should the outcomes be checked as well?

Value

A tibble containing the required predictors (and potentially the outcomes) after any required structural modifications have been made.

Details

scream() is called by forge() after shrink() but before the actual processing is done. Generally, you don't need to call scream() directly, as forge() will do it for you.

If outcomes = TRUE, then the validation steps are performed on the known outcome columns as well. If mold() was called with the XY interface, then no preprocessing was done to y and outcomes will have no effect (if a vector was passed as y during the fit, scream() has no way of knowing what column in new_data corresponds to the outcome).

If scream() is used as a standalone function, it is good practice to call shrink() right before it as there are no checks in scream() that ensure that all of the required column names actually exist in new_data. Those checks exist in shrink().

Examples

train <- iris[1:100,]
test <- iris[101:150,]

# mold() is run at model fit time
# and a terms preprocessor is recorded
x <- mold(log(Sepal.Width) ~ Species, train)

# Pass that preprocessor to shrink(), along with new_data
# to get a tibble of required predictors back
test_shrunk <- shrink(x$preprocessor, test)

# Now pass that to scream() to perform validation checks
# Silence is key!
scream(x$preprocessor, test_shrunk)

#> # A tibble: 50 x 1
#>    Species  
#>    <fct>    
#>  1 virginica
#>  2 virginica
#>  3 virginica
#>  4 virginica
#>  5 virginica
#>  6 virginica
#>  7 virginica
#>  8 virginica
#>  9 virginica
#> 10 virginica
#> # … with 40 more rows

# If `outcomes = TRUE` is used with shrink(),
# it should also be used with scream()
test_outcome <- shrink(x$preprocessor, test, outcomes = TRUE)
scream(x$preprocessor, test_outcome, outcomes = TRUE)

#> # A tibble: 50 x 2
#>    Sepal.Width Species  
#>          <dbl> <fct>    
#>  1         3.3 virginica
#>  2         2.7 virginica
#>  3         3   virginica
#>  4         2.9 virginica
#>  5         3   virginica
#>  6         3   virginica
#>  7         2.5 virginica
#>  8         2.9 virginica
#>  9         2.5 virginica
#> 10         3.6 virginica
#> # … with 40 more rows

# scream() validates that the classes of `new_data`
# are the same as the ones used in mold(). The below call
# to scream() will fail with an informative error.
test2 <- test
test2$Species <- as.character(test2$Species)
# NOT RUN {
scream(x$preprocessor, test2)
# }