Ensure that new_data contains columns of the correct class

validate - asserts the following:

The classes of the columns of new_data must be exactly the same as the classes defined by original_classes.

check - returns the following:

ok A logical. Does the check pass?

incorrect_columns A character vector. The columns that contain the wrong class.

validate_new_data_classes(new_data, original_classes)

check_new_data_classes(new_data, original_classes)

Arguments

new_data
A data frame of new predictors and possibly outcomes.

original_classes
A named list of the original classes of either the outcomes or predictors. The names should match the column names in new_data, and the values are character vectors of the original class for that column.

Details

This validation check is inspired by the base R functions stats::.MFclass() and stats::checkMFClasses(), but is more strict. See get_data_classes() for more details.

validate_new_data_classes() assumes that all of the columns in original_classes actually exist in new_data. Because of this, if you use this function by itself you should call validate_new_data_column_names() first to safely validate that the required columns actually exist.

Validation

hardhat provides validation functions at three levels.

check_*(): check a condition, and return a list. The list always contains at least one element, ok, a logical that specifies if the check passed. Each check also has check specific elements in the returned list that can be used to construct meaningful error messages.

validate_*(): check a condition, and error if it does not pass. These functions call their corresponding check function, and then provide a default error message. If you, as a developer, want a different error message, then call the check_*() function yourself, and provide your own validation function.

enforce_*(): check a condition, modify the input with a warning if the check fails, and return the modified input. These functions are used when user input is incorrect in some non-critical way, and you can correct it automatically, with a warning.

See also

Examples

# ---------------------------------------------------------------------------
# Setup

train <- iris[1:100,]
test <- iris[101:150,]

# ---------------------------------------------------------------------------
# Example usage

original_classes <- get_data_classes(train)

# All good!
check_new_data_classes(test, original_classes)

#> $ok
#> [1] TRUE
#> 
#> $incorrect_columns
#> character(0)
#> 

bad_test <- test
bad_test$Species <- as.character(bad_test$Species)

# Species is incorrect!
check_new_data_classes(bad_test, original_classes)

#> $ok
#> [1] FALSE
#> 
#> $incorrect_columns
#> [1] "Species"
#> 

# The error also tells you what the class is, and what it should be

validate_new_data_classes(bad_test, original_classes)

#> Error: Some columns in `new_data` have an incorrect class:
#> `Species`: `character` should be `factor`.