Forge prediction-ready data

forge() applies the transformations requested by the preprocessor on a set of new_data to be used in predictions.

forge(preprocessor, new_data, outcomes = FALSE, ...)

# S3 method for default_preprocessor
forge(preprocessor, new_data,
  outcomes = FALSE, ...)

# S3 method for recipes_preprocessor
forge(preprocessor, new_data,
  outcomes = FALSE, ...)

# S3 method for terms_preprocessor
forge(preprocessor, new_data,
  outcomes = FALSE, ...)

Arguments

preprocessor
A valid "preprocessor". The preprocessor that should be used here is the one in the output from the corresponding call to mold().

new_data
A data frame or matrix to preprocess.

outcomes
A logical. Should the outcomes be processed and returned as well?

...
Not currently used.

Value

A named list with elements:

predictors: A tibble containing the preprocessed new_data predictors.

outcomes: If outcomes = TRUE, the outcomes are returned here, otherwise NULL. If a formula engine was used, this is a data frame that is the result of extracting the outcome columns from stats::model.frame(). If a recipe was used, this is a data.frame that is the result of calling recipes::bake() with recipes::all_outcomes() specified.

Details

If the outcomes are present in new_data, it can optionally be processed and returned in the outcomes slot of the returned list. This is only applicable for the formula and recipes engines, but is very useful when doing cross validation where you need to preprocess the outcomes of a test set before computing performance.

Examples

# ---------------------------------------------------------------------------
# XY Method and forge(outcomes = TRUE)

# With the formula and recipes methods, you always know the name
# of the outcome columns. Because of this, you can ask for the
# outcome columns in `new_data` to be processed and returned.
# With the XY method if Y is a vector, a column name of `.outcome`
# is automatically generated. This name is what forge() looks for
# in `new_data` if `y` is a vector. If `y` is a data frame, it just
# looks for that original column name.

# X and Y are data frames
x <- iris[, "Sepal.Width", drop = FALSE]
y <- iris[, "Species", drop = FALSE]

processed <- mold(x, y)
forge(processed$preprocessor, iris, outcomes = TRUE)

#> $predictors
#> # A tibble: 150 x 1
#>    Sepal.Width
#>          <dbl>
#>  1         3.5
#>  2         3  
#>  3         3.2
#>  4         3.1
#>  5         3.6
#>  6         3.9
#>  7         3.4
#>  8         3.4
#>  9         2.9
#> 10         3.1
#> # … with 140 more rows
#> 
#> $outcomes
#> # A tibble: 150 x 1
#>    Species
#>    <fct>  
#>  1 setosa 
#>  2 setosa 
#>  3 setosa 
#>  4 setosa 
#>  5 setosa 
#>  6 setosa 
#>  7 setosa 
#>  8 setosa 
#>  9 setosa 
#> 10 setosa 
#> # … with 140 more rows
#> 

# Y is a vector
y_vec <- y$Species

processed_vec <- mold(x, y_vec)

# This throws an informative error that tell you
# to include a `".outcome"` column in `new_data`.
# NOT RUN {
forge(processed_vec$preprocessor, iris, outcomes = TRUE)
# }

iris2 <- iris
iris2$.outcome <- iris2$Species
iris2$Species <- NULL

# This works, and returns a tibble in the $outcomes slot
forge(processed_vec$preprocessor, iris2, outcomes = TRUE)

#> $predictors
#> # A tibble: 150 x 1
#>    Sepal.Width
#>          <dbl>
#>  1         3.5
#>  2         3  
#>  3         3.2
#>  4         3.1
#>  5         3.6
#>  6         3.9
#>  7         3.4
#>  8         3.4
#>  9         2.9
#> 10         3.1
#> # … with 140 more rows
#> 
#> $outcomes
#> # A tibble: 150 x 1
#>    .outcome
#>    <fct>   
#>  1 setosa  
#>  2 setosa  
#>  3 setosa  
#>  4 setosa  
#>  5 setosa  
#>  6 setosa  
#>  7 setosa  
#>  8 setosa  
#>  9 setosa  
#> 10 setosa  
#> # … with 140 more rows
#>