Construct a model frame

model_frame() is a stricter version of stats::model.frame(). There are a number of differences, with the main being that rows are never dropped and the return value is a list with the frame and terms separated into two distinct objects.

model_frame(formula, data, original_levels = NULL, drop_novel = TRUE)

Arguments

formula
A formula or terms object representing the terms of the model frame.

data
A data frame or matrix containing the terms of formula.

original_levels
Optional. A named list of character vectors. These represent the assumed factor levels for factor columns in data. If any levels are missing for a factor, then the levels are restored with a warning. If there are novel levels for a factor, then they are coerced to NA with a warning.

drop_novel
Optional. A logical. Passed on to enforce_new_data_novel_levels(), and only applicable if original_levels is not NULL. Should novel levels found in new_data but not in original_levels be coerced to NA? No matter the value of drop_novel, a warning will be thrown if any novel levels are detected.

Value

A named list with two elements:

"data": A tibble containing the model frame.

"terms": A terms object containing the terms for the model frame.

Details

The following explains the rationale for some of the difference in arguments compared to stats::model.frame():

subset: Not allowed because the number of rows before and after model_frame() has been run should always be the same.

na.action: Not allowed and is forced to "na.pass" because the number of rows before and after model_frame() has been run should always be the same.

drop.unused.levels: Not allowed because it seems inconsistent for data and the result of model_frame() to ever have the same factor column but with different levels, unless specified though original_levels. If this is required, it should be done through a recipe step explicitly.

...: Not exposed because offsets are handled separately, and it is not necessary to pass weights here any more because rows are never dropped (so weights don't have to be subset alongside the rest of the design matrix). If other non-predictor columns are required, use the "roles" features of recipes.

It is important to always use the results of model_frame() with model_matrix() rather than stats::model.matrix() because the tibble in the result of model_frame() does not have a terms object attached. If model.matrix(<terms>, <tibble>) is called directly, then a call to model.frame() will be made automatically, which can give faulty results.

Examples

# ---------------------------------------------------------------------------
# Example usage

framed <- model_frame(Species ~ Sepal.Width, iris)

framed$data

#> # A tibble: 150 x 2
#>    Species Sepal.Width
#>    <fct>         <dbl>
#>  1 setosa          3.5
#>  2 setosa          3  
#>  3 setosa          3.2
#>  4 setosa          3.1
#>  5 setosa          3.6
#>  6 setosa          3.9
#>  7 setosa          3.4
#>  8 setosa          3.4
#>  9 setosa          2.9
#> 10 setosa          3.1
#> # … with 140 more rows

framed$terms

#> Species ~ Sepal.Width
#> attr(,"variables")
#> list(Species, Sepal.Width)
#> attr(,"factors")
#>             Sepal.Width
#> Species               0
#> Sepal.Width           1
#> attr(,"term.labels")
#> [1] "Sepal.Width"
#> attr(,"order")
#> [1] 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 1
#> attr(,".Environment")
#> <environment: 0xefc2038>
#> attr(,"predvars")
#> list(Species, Sepal.Width)
#> attr(,"dataClasses")
#>     Species Sepal.Width 
#>    "factor"   "numeric" 

# ---------------------------------------------------------------------------
# Missing values never result in dropped rows

iris2 <- iris
iris2$Sepal.Width[1] <- NA

framed2 <- model_frame(Species ~ Sepal.Width, iris2)

head(framed2$data)

#> # A tibble: 6 x 2
#>   Species Sepal.Width
#>   <fct>         <dbl>
#> 1 setosa         NA  
#> 2 setosa          3  
#> 3 setosa          3.2
#> 4 setosa          3.1
#> 5 setosa          3.6
#> 6 setosa          3.9

nrow(framed2$data) == nrow(iris2)

#> [1] TRUE

# ---------------------------------------------------------------------------
# Novel and missing levels

train <- data.frame(y = factor(c("a", "b")))
test_not_enough <- data.frame(y = factor("a"))
test_too_many <- data.frame(y = factor(c("a", "b", "c")))
original_levels <- get_levels(train)

# Missing levels are recovered with a warning
model_frame(~y, test_not_enough, original_levels)

#> Warning: The following original factor levels are missing for column, 'y', and have been restored: 'b'.

#> $data
#> # A tibble: 1 x 1
#>   y    
#>   <fct>
#> 1 a    
#> 
#> $terms
#> ~y
#> attr(,"variables")
#> list(y)
#> attr(,"factors")
#>   y
#> y 1
#> attr(,"term.labels")
#> [1] "y"
#> attr(,"order")
#> [1] 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 0
#> attr(,".Environment")
#> <environment: 0xefc2038>
#> attr(,"predvars")
#> list(y)
#> attr(,"dataClasses")
#>        y 
#> "factor" 
#> 

# Novel levels are forced to NA with a warning
model_frame(~y, test_too_many, original_levels)

#> Warning: The following factor levels are new for column, `y`, and have been coerced to `NA` : 'c'.

#> $data
#> # A tibble: 3 x 1
#>   y    
#>   <fct>
#> 1 a    
#> 2 b    
#> 3 <NA> 
#> 
#> $terms
#> ~y
#> attr(,"variables")
#> list(y)
#> attr(,"factors")
#>   y
#> y 1
#> attr(,"term.labels")
#> [1] "y"
#> attr(,"order")
#> [1] 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 0
#> attr(,".Environment")
#> <environment: 0xefc2038>
#> attr(,"predvars")
#> list(y)
#> attr(,"dataClasses")
#>        y 
#> "factor" 
#> 

model_frame(~y, test_too_many, original_levels, drop_novel = FALSE)

#> Warning: The following factor levels are new for column, `y`: 'c'.

#> Warning: The following novel levels were detected in 'y': 'c'. The order of the `original_levels` has been restored, and the novel levels have been moved to the end.

#> $data
#> # A tibble: 3 x 1
#>   y    
#>   <fct>
#> 1 a    
#> 2 b    
#> 3 c    
#> 
#> $terms
#> ~y
#> attr(,"variables")
#> list(y)
#> attr(,"factors")
#>   y
#> y 1
#> attr(,"term.labels")
#> [1] "y"
#> attr(,"order")
#> [1] 1
#> attr(,"intercept")
#> [1] 1
#> attr(,"response")
#> [1] 0
#> attr(,".Environment")
#> <environment: 0xefc2038>
#> attr(,"predvars")
#> list(y)
#> attr(,"dataClasses")
#>        y 
#> "factor" 
#>