Mold data for modeling

mold() applies the appropriate processing steps required to get training data ready to be fed into a model.

For a formula, this applies stats::model.frame() and possibly stats::model.matrix().

For a recipe, this performs a call to both recipes::prep() and recipes::juice().

For a data frame or matrix, this simply adds an intercept column if requested.

mold(x, ...)

# S3 method for data.frame
mold(x, y, intercept = FALSE, ...)

# S3 method for matrix
mold(x, y, intercept = FALSE, ...)

# S3 method for formula
mold(formula, data, intercept = FALSE,
  indicators = TRUE, ...)

# S3 method for recipe
mold(x, data, intercept = FALSE, ...)

Arguments

x
A data frame, matrix, or recipes::recipe(). If this is a data.frame or matrix, it should contain the predictors.

...
Currently unused.

y
A data frame, matrix, or vector containing the outcome(s).

intercept
A single logical specifying whether or not to include an intercept in the molded predictors.

formula
A formula specifying the terms in the format of outcome ~ predictors.

data
A data frame containing the predictors and the outcomes.

indicators
For use with the formula interface. Should factors and interactions be expanded (In other words, should model.matrix() be run)? If FALSE, factor columns are returned without being expanded into dummy variables and a warning is thrown if any interactions are detected.

Value

A named list containing:

predictors: A tibble containing the molded predictors to be used in the model.

outcome: A tibble.

If y was supplied, it is returned after a call to standardize() is made.

If a formula engine was used, this is a data frame that is the result of extracting the molded outcome columns from model.frame().

If a recipe was used, this is a data.frame that is the result of calling recipes::juice() with recipes::all_outcomes() specified.

preprocessor: A "preprocessor" object for use when making predictions.

Details

Multivariate outcomes can be specified on the LHS using similar syntax as the RHS (i.e. outcome_1 + outcome_2 ~ predictors). stats::model.frame() is run on the outcome columns (which will execute any in line formulas), but stats::model.matrix() is not run on the outcomes as this would expand factor outcomes and this is likely not desired. If any complex calculations are done on the LHS and they return matrices (like stats::poly()), then those matrices are flattened into multiple columns of the tibble after the call to model.frame().

Examples

# ---------------------------------------------------------------------------
# Multivariate outcomes

# Multivariate formulas can be specified easily
processed <- mold(Sepal.Width + log(Sepal.Length) ~ Species, iris)
processed$outcomes

#> # A tibble: 150 x 2
#>    Sepal.Width `log(Sepal.Length)`
#>          <dbl>               <dbl>
#>  1         3.5                1.63
#>  2         3                  1.59
#>  3         3.2                1.55
#>  4         3.1                1.53
#>  5         3.6                1.61
#>  6         3.9                1.69
#>  7         3.4                1.53
#>  8         3.4                1.61
#>  9         2.9                1.48
#> 10         3.1                1.59
#> # … with 140 more rows

# Inline functions on the LHS are run, but any matrix
# output is flattened (like what happens in `model.matrix()`)
# (essentially this means you don't wind up with columns
# in the tibble that are matrices)
processed <- mold(poly(Sepal.Length, degree = 2) ~ Species, iris)
processed$outcomes

#> # A tibble: 150 x 2
#>    `poly(Sepal.Length, degree = 2).1` `poly(Sepal.Length, degree = 2).2`
#>                                 <dbl>                              <dbl>
#>  1                            -0.0735                            0.00652
#>  2                            -0.0933                            0.0469 
#>  3                            -0.113                             0.0957 
#>  4                            -0.123                             0.123  
#>  5                            -0.0834                            0.0257 
#>  6                            -0.0439                           -0.0385 
#>  7                            -0.123                             0.123  
#>  8                            -0.0834                            0.0257 
#>  9                            -0.143                             0.184  
#> 10                            -0.0933                            0.0469 
#> # … with 140 more rows

# TRUE
ncol(processed$outcomes) == 2

#> [1] TRUE