Mold data for modeling

mold() applies the appropriate processing steps required to get training data ready to be fed into a model.

For a formula, this applies stats::model.frame() and possibly stats::model.matrix().

For a recipe, this performs a call to both recipes::prep() and recipes::juice().

For a data frame or matrix, this uses the new_default_preprocessor() which converts the input to an object of class type and adds an intercept column if requested.

mold(x, ...)

# S3 method for data.frame
mold(x, y, intercept = FALSE, type = "tibble",
  ...)

# S3 method for matrix
mold(x, y, intercept = FALSE, type = "tibble", ...)

# S3 method for formula
mold(formula, data, intercept = FALSE,
  type = "tibble", indicators = TRUE, ...)

# S3 method for recipe
mold(x, data, intercept = FALSE, type = "tibble", ...)

Arguments

x
A data frame, matrix, or recipes::recipe(). If this is a data.frame or matrix, it should contain the predictors.

...
Currently unused.

y
A data frame, matrix, or vector containing the outcome(s).

intercept
A single logical specifying whether or not to include an intercept in the molded predictors.

type
A single character. One of "tibble", "data.frame", or "matrix" specifying the result type of the predictors.

formula
A formula specifying the terms in the format of outcome ~ predictors.

data
A data frame containing the predictors and the outcomes.

indicators
For use with the formula interface. Should factors and interactions be expanded (In other words, should model.matrix() be run)? If FALSE, factor columns are returned without being expanded into dummy variables and a warning is thrown if any interactions are detected.

Value

A named list containing:

predictors: An object of class type containing the molded predictors to be used in the model.

outcome: A tibble.

If y was supplied, it is returned after a call to standardize() is made.

If a formula engine was used, this is a data frame that is the result of extracting the molded outcome columns from model.frame().

If a recipe was used, this is a data.frame that is the result of calling recipes::juice() with recipes::all_outcomes() specified.

preprocessor: A "preprocessor" object for use when making predictions.