Subset only required columns

shrink() subsets new_data to only contain the required columns. shrink() also performs the following validation checks on new_data:

validate_new_data_column_names() - Checks that all required columns exist in new_data.

shrink(preprocessor, new_data, outcomes = FALSE)

Arguments

preprocessor
A "preprocessor".

new_data
A data frame at least containing the new predictors, and potentially the outcomes if outcomes = TRUE.

outcomes
A logical. Should the outcomes be included in the shrunk new_data output as well?

Value

A tibble containing the required predictors (and potentially the outcomes).

Details

If outcomes = TRUE, then the validation checks are performed on the known outcome columns as well, and they are returned along with the predictors. If mold() was called with the XY interface, then no preprocessing was done to y and outcomes will have no effect (if a vector was passed as y during the fit, shrink() has no way of knowing what column in new_data corresponds to the outcome).

shrink() is called by forge() before scream() and before the actual processing is done.

Examples

train <- iris[1:100,]
test <- iris[101:150,]

# mold() is run at model fit time
# and a terms preprocessor is recorded
x <- mold(log(Sepal.Width) ~ Species, train)

# Pass that preprocessor to shrink(), along with new_data
# to get a tibble of required predictors back
shrink(x$preprocessor, test)

#> # A tibble: 50 x 1
#>    Species  
#>    <fct>    
#>  1 virginica
#>  2 virginica
#>  3 virginica
#>  4 virginica
#>  5 virginica
#>  6 virginica
#>  7 virginica
#>  8 virginica
#>  9 virginica
#> 10 virginica
#> # … with 40 more rows

# outcomes = TRUE also returns the outcomes,
# this can be useful if doing cross validation
shrink(x$preprocessor, test, outcomes = TRUE)

#> # A tibble: 50 x 2
#>    Sepal.Width Species  
#>          <dbl> <fct>    
#>  1         3.3 virginica
#>  2         2.7 virginica
#>  3         3   virginica
#>  4         2.9 virginica
#>  5         3   virginica
#>  6         3   virginica
#>  7         2.5 virginica
#>  8         2.9 virginica
#>  9         2.5 virginica
#> 10         3.6 virginica
#> # … with 40 more rows

# forge() actually preprocesses the new_data, after
# a call to shrink(). Notice how now the Sepal.Width
# column is logged
forge(x$preprocessor, test, outcomes = TRUE)$outcomes

#> # A tibble: 50 x 1
#>    `log(Sepal.Width)`
#>                 <dbl>
#>  1              1.19 
#>  2              0.993
#>  3              1.10 
#>  4              1.06 
#>  5              1.10 
#>  6              1.10 
#>  7              0.916
#>  8              1.06 
#>  9              0.916
#> 10              1.28 
#> # … with 40 more rows