Subset only required columns

shrink() performs a number of validation steps on new_data, and then subsets it to only contain the required columns. shrink() does the following validation:

Checks that all required predictor columns exist in new_data.

Checks that the class of each required predictor in new_data is the same as the class used during training.

Checks that all new_data factor columns don't have any new levels when compared with the original data used in training. If there are new levels, they are replaced with NA values and a warning is issued.

Checks that all new_data factor columns aren't missing any factor levels when compared with the original data used in training. If there are missing levels (or misordered levels for ordered factors), then they are restored and a warning is issued.

shrink(preprocessor, new_data, outcome = FALSE)

Arguments

preprocessor
A "preprocessor".

new_data
A data frame containing the new data that will be shrunk down to only contain the predictors required for preprocessing and prediction.

outcome
A logical. Should the outcome be included in the shrunk new_data as well?

Value

A tibble containing the required predictors (and potentially the outcomes).

Details

If outcome = TRUE, then the validation steps are performed on the known outcome column as well, and it is returned along with the predictors. If mold() was called with the XY interface, then no preprocessing was done to y and outcome cannot be set to TRUE (if a vector was passed as y during the fit, shrink() has no way of knowing what column in new_data corresponds to the outcome).

shrink() is called by forge() before the actual processing is done.

Examples

train <- iris[1:100,]
test <- iris[101:150,]

# mold() is run at model fit time
# and a terms preprocessor is recorded
x <- mold(log(Sepal.Width) ~ Species, train)

# Pass that preprocessor to shrink(), along with new_data
# to get a tibble of required predictors back
shrink(x$preprocessor, test)

#> # A tibble: 50 x 1
#>    Species  
#>    <fct>    
#>  1 virginica
#>  2 virginica
#>  3 virginica
#>  4 virginica
#>  5 virginica
#>  6 virginica
#>  7 virginica
#>  8 virginica
#>  9 virginica
#> 10 virginica
#> # … with 40 more rows

# outcome = TRUE also returns the outcome,
# this can be useful if doing cross validation
shrink(x$preprocessor, test, outcome = TRUE)

#> # A tibble: 50 x 2
#>    Sepal.Width Species  
#>          <dbl> <fct>    
#>  1         3.3 virginica
#>  2         2.7 virginica
#>  3         3   virginica
#>  4         2.9 virginica
#>  5         3   virginica
#>  6         3   virginica
#>  7         2.5 virginica
#>  8         2.9 virginica
#>  9         2.5 virginica
#> 10         3.6 virginica
#> # … with 40 more rows

# shrink() validates that the classes are the same
# as the ones used in mold(). The below call
# to shrink() will fail with an informative error.
test2 <- test
test2$Species <- as.character(test2$Species)
# NOT RUN {
shrink(x$preprocessor, test2)
# }